+++ sim/integration/v1_solver_mono_wide_trapezoidal_tb.sv
@@ -10,7 +10,8 @@ module v1_solver_mono_wide_trapezoidal_tb #(
     parameter bit PIPELINED_KCL_ACCUMULATOR = 1'b0,
     parameter bit PIPELINED_KCL_CAPACITOR_CURRENT = 1'b0,
     parameter bit PIPELINED_KCL_MAXIMUM = 1'b0,
-    parameter bit PIPELINED_CHORD_APPLY = 1'b0
+    parameter bit PIPELINED_CHORD_APPLY = 1'b0,
+    parameter bit HALF_PARALLEL_TERMINAL_CURRENT = 1'b0
 );
     v1_solver_mono_wide_tb #(
         .TRAPEZOIDAL(1'b1),
@@ -22,7 +23,10 @@ module v1_solver_mono_wide_trapezoidal_tb #(
         .PIPELINED_KCL_ACCUMULATOR(PIPELINED_KCL_ACCUMULATOR),
         .PIPELINED_KCL_CAPACITOR_CURRENT(PIPELINED_KCL_CAPACITOR_CURRENT),
         .PIPELINED_KCL_MAXIMUM(PIPELINED_KCL_MAXIMUM),
-        .PIPELINED_CHORD_APPLY(PIPELINED_CHORD_APPLY)
+        .PIPELINED_CHORD_APPLY(PIPELINED_CHORD_APPLY),
+        .HALF_PARALLEL_TERMINAL_CURRENT(
+            HALF_PARALLEL_TERMINAL_CURRENT
+        )
     ) testbench();
 endmodule